pyproject.toml
[project]
name = "image-classification-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "jinja2>=3.1.6",
    "pillow>=11.3.0",
    "python-multipart>=0.0.20",
    "tensorflow>=2.20.0",
    "unicorn>=2.1.3",
    "uvicorn>=0.35.0",
]
